feat: Set up a Flask server
Since working with Streamlit did not yield workable results, a
Flask server has been set up to handle the user input and the
output of the game.

A basic frontend has been created.

Flask triggers the beginning of the game upon loading the page and
displays the introduction via Jinja. The user is prompted for input.

The input is directed by a JavaScript script, to prevent Flask from
rendering the page without the necessary output while the game
progresses.

Flask then receives the updated game, but fails to display it in the
browser dynamically.