feat: Adjust the game to be displayed in the browser
The main game script, along with a class method, have been adjusted
to work with the server to serve the game in the browser, instead of
printing the output in the console.

The player input is now also received from the browser instead of from
the terminal.

The player does not have to scroll to the bottom of the text box to
read the latest game update. The text box scrolls automatically to
the bottom.