feat: Style the UI
A background for the game UI has been added, as well as the logo.

The text box and input box have been styled to match the theme,
using the brand colors.

The font has been added and adjusted for better readability while
matching the website's theme.

The scrollbar has been styled to match the theme across all
browsers.