feat: Integrate Staff of Elementia into Coding Sorceress
Staff of Elementia can now be run from inside of Coding Sorceress.

A header with a link back to Coding Sorceress has been added to the
UI of Staff of Elementia for the user's convenience.

The access to Staff od Elementia has been secured against users who
are not logged in into Coding Sorceress.